remove extar js file as a must
