+++ src/script.js
@@ -1,15 +1,9 @@
 import './style.css';
-import { displayTasks } from './interactive.js';
 import * as crud from './CRUD.js';
+import { tasks, todoList, displayTasks } from './CRUD.js';
 
-export const form = document.getElementById('form');
-export const input = document.getElementById('your-todo');
-export const todoList = document.getElementById('todo-list');
 const clearAll = document.getElementById('clear-all');
-export const refMe = document.querySelector('.bi-arrow-clockwise');
 const enterIcon = document.querySelector('.bi-box-arrow-in-left');
-
-export const tasks = JSON.parse(localStorage.getItem('tasks')) || [];
 
 function clearAllFunc() {
   const task = tasks.filter((task) => !task.completed);
@@ -21,7 +15,7 @@ function clearAllFunc() {
   window.location.reload();
 }
 
-form.addEventListener('submit', crud.addItem);
+crud.form.addEventListener('submit', crud.addItem);
 todoList.addEventListener('click', crud.toggleItem);
 enterIcon.addEventListener('click', crud.addItem);
 clearAll.addEventListener('click', clearAllFunc);
